+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 - [#827](https://github.com/hyperf/hyperf/pull/827) Added a simple db component.
 - [#905](https://github.com/hyperf/hyperf/pull/905) Added twig template engine for view.
+- [#911](https://github.com/hyperf/hyperf/pull/911) Added support for crontab task run on one server.
 - [#913](https://github.com/hyperf/hyperf/pull/913) Added `Hyperf\ExceptionHandler\Listener\ErrorExceptionHandler`.
+- [#931](https://github.com/hyperf/hyperf/pull/931) Added `strict_mode` for config-apollo.
 
 ## Fixed